requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
dateparser==1.0.0
dj-database-url==0.5.0
Django==2.2.13
django-bootstrap4==3.0.0
django-heroku==0.3.1
gunicorn==20.1.0
idna==2.10
importlib-metadata==2.1.1
multidict==5.1.0
numpy==1.20.2
pandas==1.2.4
psycopg2==2.8.6
python-binance==1.0.2
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
six==1.10.0
sortedcontainers==2.3.0
soupsieve==2.2.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.4
websocket-client==0.40.0
websockets==9.0.1
whitenoise==5.2.0
yarl==1.6.3
zipp==3.4.1
